--- src/app/gpt/auth.py
@@ -1,9 +1,0 @@
-# import typing as ty
-
-# from src.app.actor import System
-# from src.app.gpt.model import CreateUser, User, UserCreated
-
-
-# class UserSystem(System[ty.Any]):
-#     def create_user(self, command: CreateUser) -> User:
-#         ...
