+++ src/app/gpt/repository.py
@@ -4,7 +4,7 @@ import typing as ty
 import sqlalchemy as sa
 from sqlalchemy.ext import asyncio as sa_aio
 
-from src.app.gpt.model import IUserRepository, User
+from src.app.gpt.model import ISessionRepository, IUserRepository, User
 
 
 def user_deserializer(user_data: dict[str, ty.Any]) -> User:
@@ -30,9 +30,14 @@ class UserRepository(IUserRepository):
 
     async def search_user_by_email(self, useremail: str) -> User | None:
         sql = """
-        SELECT * from domain_events WHERE event_type = :event_type and event_body->>'user_info'->> 'user_email' = :useremail
+        SELECT 
+            * 
+        FROM 
+            domain_events 
+        WHERE 
+            event_body->>'user_info'->>'user_email' = :useremail
         """
-        stmt = sa.text(sql).bindparams(event_type="user_created", useremail=useremail)
+        stmt = sa.text(sql).bindparams(useremail=useremail)
 
         async with self.aioengine.begin() as conn:
             cursor = await conn.execute(stmt)
@@ -43,3 +48,8 @@ class UserRepository(IUserRepository):
 
         user_data = dict(res._mapping)  # type: ignore
         return user_deserializer(user_data)
+
+
+class SessionRepository(ISessionRepository):
+    def __init__(self, aioengine: sa_aio.AsyncEngine):
+        self._aioengine = aioengine
