+++ tests/e2e/conftest.py
@@ -4,6 +4,7 @@ import pathlib
 import pytest
 
 from src.domain.config import Settings
+from src.domain.fileutil import fileutil
 
 
 @pytest.fixture(scope="session")
@@ -17,7 +18,7 @@ class TestSettings(Settings):
     __test__ = False
 
     class DB(Settings.DB):
-        DATABASE: pathlib.Path  # = pathlib.Path(":memory:")
+        DATABASE: pathlib.Path
         ENGINE_ECHO: bool = False
 
     class ActorRefs(Settings.ActorRefs):
@@ -27,9 +28,13 @@ class TestSettings(Settings):
 @pytest.fixture(scope="session")
 def settings() -> TestSettings:
     db_path = pathlib.Path("./database/test.db")
+    # db_path = pathlib.Path(":memory:")
+    # api_key = fileutil.read_file("settings.toml")["OPENAI_API_KEY"]
+    api_key = "fake_api_key"
+
     db = TestSettings.DB(DATABASE=db_path)
     ss = TestSettings(
-        OPENAI_API_KEY="fake_api_key",
+        OPENAI_API_KEY=api_key,
         db=db,
         actor_refs=TestSettings.ActorRefs(),
         RUNTIME_ENV="test",
